+++ stubs/inertia-vue-ts/resources/js/Components/TextInput.vue
@@ -1,11 +1,7 @@
 <script setup lang="ts">
 import { onMounted, ref } from 'vue';
 
-defineProps<{
-    modelValue: string;
-}>();
-
-defineEmits(['update:modelValue']);
+const model = defineModel<string>({ required: true });
 
 const input = ref<HTMLInputElement | null>(null);
 
@@ -21,8 +17,7 @@ defineExpose({ focus: () => input.value?.focus() });
 <template>
     <input
         class="border-gray-300 dark:border-gray-700 dark:bg-gray-900 dark:text-gray-300 focus:border-indigo-500 dark:focus:border-indigo-600 focus:ring-indigo-500 dark:focus:ring-indigo-600 rounded-md shadow-sm"
-        :value="modelValue"
-        @input="$emit('update:modelValue', ($event.target as HTMLInputElement).value)"
+        v-model="model"
         ref="input"
     />
 </template>