+++ stubs/inertia-vue/resources/js/Components/TextInput.vue
@@ -1,14 +1,10 @@
 <script setup>
 import { onMounted, ref } from 'vue';
 
-defineProps({
-    modelValue: {
-        type: String,
-        required: true,
-    },
+const model = defineModel({
+    type: String,
+    required: true,
 });
-
-defineEmits(['update:modelValue']);
 
 const input = ref(null);
 
@@ -24,8 +20,7 @@ defineExpose({ focus: () => input.value.focus() });
 <template>
     <input
         class="border-gray-300 dark:border-gray-700 dark:bg-gray-900 dark:text-gray-300 focus:border-indigo-500 dark:focus:border-indigo-600 focus:ring-indigo-500 dark:focus:ring-indigo-600 rounded-md shadow-sm"
-        :value="modelValue"
-        @input="$emit('update:modelValue', $event.target.value)"
+        v-model="model"
         ref="input"
     />
 </template>