+++ stubs/livewire-functional/resources/views/livewire/pages/auth/login.blade.php
@@ -17,10 +17,7 @@ $login = function () {
 
     Session::regenerate();
 
-    $this->redirect(
-        session('url.intended', route('dashboard', absolute: false)),
-        navigate: true
-    );
+    $this->redirectIntended(default: route('dashboard', absolute: false), navigate: true);
 };
 
 ?>