+++ stubs/livewire-functional/resources/views/livewire/pages/auth/verify-email.blade.php
@@ -10,10 +10,7 @@ layout('layouts.guest');
 
 $sendVerification = function () {
     if (Auth::user()->hasVerifiedEmail()) {
-        $this->redirect(
-            session('url.intended', route('dashboard', absolute: false)),
-            navigate: true
-        );
+        $this->redirectIntended(default: route('dashboard', absolute: false), navigate: true);
 
         return;
     }