+++ stubs/livewire-functional/resources/views/livewire/profile/update-profile-information-form.blade.php
@@ -36,9 +36,7 @@ $sendVerification = function () {
     $user = Auth::user();
 
     if ($user->hasVerifiedEmail()) {
-        $path = session('url.intended', route('dashboard', absolute: false));
-
-        $this->redirect($path);
+        $this->redirectIntended(default: route('dashboard', absolute: false));
 
         return;
     }