+++ stubs/livewire/resources/views/livewire/pages/auth/confirm-password.blade.php
@@ -29,10 +29,7 @@ new #[Layout('layouts.guest')] class extends Component
 
         session(['auth.password_confirmed_at' => time()]);
 
-        $this->redirect(
-            session('url.intended', route('dashboard', absolute: false)),
-            navigate: true
-        );
+        $this->redirectIntended(default: route('dashboard', absolute: false), navigate: true);
     }
 }; ?>