+++ stubs/livewire/resources/views/livewire/pages/auth/login.blade.php
@@ -20,10 +20,7 @@ new #[Layout('layouts.guest')] class extends Component
 
         Session::regenerate();
 
-        $this->redirect(
-            session('url.intended', route('dashboard', absolute: false)),
-            navigate: true
-        );
+        $this->redirectIntended(default: route('dashboard', absolute: false), navigate: true);
     }
 }; ?>