+++ stubs/livewire/resources/views/livewire/pages/auth/verify-email.blade.php
@@ -14,10 +14,7 @@ new #[Layout('layouts.guest')] class extends Component
     public function sendVerification(): void
     {
         if (Auth::user()->hasVerifiedEmail()) {
-            $this->redirect(
-                session('url.intended', route('dashboard', absolute: false)),
-                navigate: true
-            );
+            $this->redirectIntended(default: route('dashboard', absolute: false), navigate: true);
 
             return;
         }