+++ stubs/livewire/resources/views/livewire/profile/update-profile-information-form.blade.php
@@ -51,9 +51,7 @@ new class extends Component
         $user = Auth::user();
 
         if ($user->hasVerifiedEmail()) {
-            $path = session('url.intended', route('dashboard', absolute: false));
-
-            $this->redirect($path);
+            $this->redirectIntended(default: route('dashboard', absolute: false));
 
             return;
         }